+++ src/duplo_resource/ecs_service.py
@@ -1,6 +1,6 @@
 from duplocloud.client import DuploClient
 from duplocloud.resource import DuploTenantResourceV2
-from duplocloud.errors import DuploError
+from duplocloud.errors import DuploError, DuploStillWaiting
 from duplocloud.commander import Command, Resource
 import duplocloud.args as args
 
@@ -277,4 +277,4 @@ class DuploEcsService(DuploTenantResourceV2):
     # filter the tasks down to any where the DesiredStatus and LastStatus are different
     running_tasks = [t for t in tasks if t["DesiredStatus"] != t["LastStatus"]]
     if len(running_tasks) > 0:
-      raise DuploError(f"Service {name} waiting for replicas update", 400)
+      raise DuploStillWaiting(f"Service {name} waiting for replicas update")