+++ src/duplo_resource/hosts.py
@@ -1,6 +1,6 @@
 from duplocloud.client import DuploClient
 from duplocloud.resource import DuploTenantResourceV2
-from duplocloud.errors import DuploError, DuploFailedResource
+from duplocloud.errors import DuploError, DuploFailedResource, DuploStillWaiting
 from duplocloud.commander import Command, Resource
 import duplocloud.args as args
 
@@ -63,7 +63,7 @@ class DuploHosts(DuploTenantResourceV2):
       if h["Status"] != "running":
         if h["Status"] != "pending":
           raise DuploFailedResource(f"Host '{name}' failed to create.")
-        raise DuploError(f"Host '{name}' not ready", 404)
+        raise DuploStillWaiting(f"Host '{name}' is waiting")
     # let's get started
     if body.get("ImageId", None) is None:
       body["ImageId"] = self.discover_image(body.get("AgentPlatform", 0))
@@ -113,7 +113,7 @@ class DuploHosts(DuploTenantResourceV2):
         else:
           raise DuploFailedResource(f"Host '{name}' failed to delete.")
       if h["Status"] == "shutting-down" or h["Status"] == "running":
-        raise DuploError(f"Host '{name}' not terminated", 404)
+        raise DuploStillWaiting(f"Host '{name}' is waiting for termination")
     if wait:
       self.wait(wait_check, 500)
     return {
@@ -156,7 +156,7 @@ class DuploHosts(DuploTenantResourceV2):
       if h["Status"] != "stopped":
         if h["Status"] != "stopping":
           raise DuploFailedResource(f"Host '{name}' failed to stop.")
-        raise DuploError(f"Host '{name}' not ready", 404)
+        raise DuploStillWaiting(f"Host '{name}' is waiting for status stopped")
     if wait:
       self.wait(wait_check, 500)
     return {
@@ -199,7 +199,7 @@ class DuploHosts(DuploTenantResourceV2):
       if h["Status"] != "running":
         if h["Status"] != "pending":
           raise DuploFailedResource(f"Host '{name}' failed to stop.")
-        raise DuploError(f"Host '{name}' not ready", 404)
+        raise DuploStillWaiting(f"Host '{name}' is waiting for status running")
     if wait:
       self.wait(wait_check, 500)
     return {